requirements.txt
Flask==3.0.2
Jinja2==3.1.3
python-dotenv==1.0.1
requests==2.31.0